Research Internship on RISC-V using VSDSquadron Mini RISC-V Dev Board
Thrilled to embark on the VSDSquadron Mini RISC-V internship, I look forward to a hands-on exploration of this cutting-edge architecture. With assured resources and support, the four-week program promises an immersive learning experience. Excited to apply theoretical knowledge to practical projects, I anticipate challenges and growth. Eager to unravel Mini RISC-V's intricacies, I'm ready for an enjoyable journey of discovery and skill enhancement.